moved styled components to one file
