+++ src/Context/Store.js
@@ -10,7 +10,7 @@ export const VerbContext = React.createContext([{}, () => {}]);
 
 const Store = ({ children }) => {
   const [colorFill, setColorFill] = useState("#fff");
-  const [modal, setmodal] = useState(false);
+  const [modal, setModal] = useState(false);
   const [register, setRegister] = useState(false);
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
@@ -19,7 +19,7 @@ const Store = ({ children }) => {
 
   return (
     <ColorFillContext.Provider value={[colorFill, setColorFill]}>
-      <ModalContext.Provider value={[modal, setmodal]}>
+      <ModalContext.Provider value={[modal, setModal]}>
         <UsernameContext.Provider value={[username, setUsername]}>
           <PasswordContext.Provider value={[password, setPassword]}>
             <RegisterContext.Provider value={[register, setRegister]}>
