+++ src/app.css
@@ -15,7 +15,6 @@ body {
   margin: 0 auto;
   border: 1px solid green;
 }
-
 /* Desktop styles */
 
 @media only screen and (min-width: 630px) {
